+++ src/main/java/br/com/desafiosolutis/repository/PautaRepository.java
@@ -6,11 +6,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.desafiosolutis.model.Pauta;
 
-public interface PautaRepository extends JpaRepository<Pauta, Long> {
+public interface PautaRepository extends JpaRepository<Pauta, Integer> {
 
 
 
 	List<Pauta> findByNomePauta(String nomePauta);
 
-	boolean existsById(Long id);
+	boolean existsById(Integer id);
 }
